fix(qcbor): Resolve -Wcast-qual warnings in .c files
Continues patching the 3rd-party QCBOR library to compile
with stricter flags.

- nm/contrib/QCBOR/src/UsefulBuf.c:
  Changes several pointer casts from (uint8_t *) to
  (const uint8_t *) to maintain const-correctness during
  pointer arithmetic.

- nm/contrib/QBORN/src/qcbor_decode.c:
  Updates the UNCONST_POINTER macro to use pointer-laundering
  ((void *)(uintptr_t)(ptr)). This preserves its explicit
  intent to discard 'const' in a standards-compliant way.

+++ nm/contrib/QCBOR/src/qcbor_decode.c
@@ -71,7 +71,7 @@ IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  This casts away the const-ness of a pointer, usually so it can be
  freed or realloced.
  */
-#define UNCONST_POINTER(ptr)    ((void *)(ptr))
+#define UNCONST_POINTER(ptr)    ((void *)(uintptr_t)(ptr))
 
 
 /*
